#15 - Optimize and refactor the folder structure
