+++ practice-one/src/components/DataDisplay/Drawer/DrawerItem/DrawerItem.scss
@@ -1,4 +1,4 @@
-.sidebar__item {
+.drawer__item {
   display         : flex;
   align-items     : center;
   background-color: #3b3e46;
