+++ practice-one/src/components/DataDisplay/index.ts
@@ -1,0 +1,1 @@
+export { default as Drawer } from './Drawer/index';
