+++ practice-one/src/components/Inputs/Button/Button.scss
@@ -28,7 +28,7 @@
     padding: 8px 40px;
   }
 
-  &__sidebar {
+  &__drawer {
     border: 1px solid $primary-color;
     margin: 18px 0 0 20px
   }
