+++ practice-one/src/components/Inputs/SearchBar/SearchBar.scss
