+++ practice-one/src/components/Inputs/index.ts
@@ -1,5 +1,3 @@
 export { default as Button } from './Button/index';
-export { default as Heading } from './Heading/index';
-export { default as Sidebar } from './Sidebar/index';
 export { default as SearchBar } from './SearchBar/index';
 export { default as TextField } from './TextField/index';
