--- practice-one/src/components/commons/Heading/heading.scss
@@ -1,3 +1,0 @@
-.header__heading {
-  line-height: calc(32/16);
-}
