+++ practice-one/src/styles/components/index.scss
@@ -1,6 +1,1 @@
-@import './form.scss';
-@import'./modal.scss';
-@import'./overlay.scss';
-@import './table.scss';
 @import './typography.scss';
-@import './validation.scss';
