+++ practice-one/src/styles/pages/index.scss
@@ -1,3 +1,1 @@
 @import './home/header.scss';
-@import './home/user-details.scss';
-@import './home/edit-user.scss';
